requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2

redis[hiredis]==5.2.1

python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
httpx==0.28.1
sse-starlette==2.2.1
